Implement wide array optimisation
If all accesses of a given array are of its element type, and each
offset through which it is accessed is known to be a multiple of the
element size, model that array as a map from offsets to elements,
rather than from offsets to bytes.

The translator begins by assuming that all arrays may be modelled
using elements.  If it encounters any arrays which cannot be modelled
in this way, it will restart the translation process with those arrays
modelled using bytes.