Pointer analysis based on fixpoints
This implements a form of pointer analysis which reuses the
translator's existing fixpoint mechanism for deciding which arrays to
model as byte arrays.  If all callsites for a function pass an offset
into a particular array as a particular parameter, that parameter is
modelled as an offset into that array.