Prevent VarRefExprs from being assigned SSA variables
Normally the scope of an SSA variable is the same as that of the
corresponding value in LLVM IR (i.e. the owning basic block and
its dominees).  In cases where Bugle can simplify the LLVM value to
a subexpression, it ought to use the subexpression's SSA variable
(or none at all, for example if the subexpression was derived from a
constant) rather than assigning a new variable to the subexpression,
which may not have sufficient scope for any uses of the subexpression
elsewhere.  But there was (mistakenly) nothing preventing Bugle from
assigning an SSA variable to the subexpression if it was a VarRefExpr.
This change fixes that.