Replace DenseMaps by std::maps wholesale
Some of the DenseMaps were used in recursive
algorithms, with references to elements in the
map held higher up the stack. When a DenseMap is
resized, all the held references become invalid.
This was not taken into account; leading to
memory corruption. std::map references stay valid
on the other hand.