Provide axioms for unconstrained dimensions
The tricky one is where we have a fixed global
size but an unconstrained local size. For the
others an an axiom stating that the unconstrained
value is greater than 0 suffices. It could be
argued that in this latter case the axiom could
be omitted, but then relevant variable is not
generated, which might lead to problems in the
invariant generation code.