Fix issue where invalid array bound was generated
Due to the order in which functions are traversed and in which
global arrays are added, it is not always the case that an
array passed to a kernel entry pointed is recorded as being
such. Correct this later when we reach the entry point.

Note that explicit bounds passed on the command line will only
be assigned after recording that an array is coming from an
entry point. Hence, temporarily invalidating the 0-th dimension
(which is what happens when an entry point is reached), is not
an issue.