Obtain the function argument names from the debug information
Current LLVM release builds generate nameless function arguments,
while before the function arguments used to derive from the names
they carried in the source code. Hence, the function arguments in
the bitcode are no longer of use to us, and we need to resort to
the debug information. Somewhat annoyingly the relevant debug
information occurs in the entry block of the function. Hence, we
need to drill down in the block to retrieve it.