Handle extract element instructions with non-constant indexes
The instructions are handled by generating a case distinction
over all possible indexes. The case distinction consists of a
number of nested if-then-else expressions.

For the case where the index is constant we still generates the
same Boogie code as before, but this is now a consequence of the
constant folding that is built into the expression creation
methods.