Simplify memcpy usages during preprocessing to make mem2reg applicable
This reduces the number of kernels in the invariant generation benchmark
set that alloca structs to 16. These 16 can be divided in groups of 2~3
kernels each, with each group needing different preprocessing to remove the
remaing allocas of structs. Given the sizes of these groups, this currently
does not seem to warrant further investment of effort.